Closes #2712 Adds Media Tag Carry Over on Duplicate
# Issue #2712

# Summary
Creates function for duplicating media into another occurrence. Replaces
existing logic to consolidate it in the Media class with the added
functionality of pull the media tags over as well.

It should be noted that this added functionality requires the existing
single query to get broken into a bunch of smaller queries which may
impact performance, likely not enought to take notice.